requirements.txt

# Core project dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
loguru

# Web UI
streamlit>=1.28.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
